db_server/requirements.txt
aiofiles==25.1.0
Pillow>=11.0.0
fastapi==0.135.2
mysql-connector-python==9.6.0
pymcprotocol==0.3.0
python-multipart==0.0.22
uvicorn==0.42.0
websockets==16.0
